Refactor code to be more concise and efficient
Hoist check_can_access_instance code into bmcs_helper.rb.
Annotate each of the per-vnic info with a vnic# tag in output.
